move fuser-specific parts into fuser bin
